Updated windows make options and added gitignore file
Added a gitignore file to ignore the object file, save and template
files, and other generated-while-playing type resources, as well as
eclipse ide related resources.
Also committing the updated windows makefile which has been modified
slightly for development work - updated the executable to run to g++,
enabled debugging and set it to extended-mode for gdb debugging, and
disabled optimisation.

Optimisation is intended for final release builds and whilst it is
certainly worth testing, it does sometimes interfere with debugging in
dev builds.